✨ Move header and footer to app.vue
