+++ src/router/index.ts
@@ -1,23 +1,16 @@
 import { createRouter, createWebHashHistory } from 'vue-router';
 import type { RouteRecordRaw } from 'vue-router';
-import App from '../app.vue';
 
 const routes: RouteRecordRaw[] = [
   {
     path: '/',
-    component: App,
-    children: [
-      {
-        path: '',
-        name: 'Timers',
-        component: () => import('../views/Index.vue')
-      },
-      {
-        path: 'reports',
-        name: 'Reports',
-        component: () => import('../views/Reports.vue')
-      }
-    ]
+    name: 'Timers',
+    component: () => import('../views/Index.vue')
+  },
+  {
+    path: '/reports',
+    name: 'Reports',
+    component: () => import('../views/Reports.vue')
   },
   // Add a catch-all route that redirects to the home page
   {
